requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.2
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httplib2==0.22.0
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
langchain-core==0.1.46
langchain-google-genai==1.0.3
langsmith==0.1.51
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest_asyncio==1.6.0
orjson==3.10.1
packaging==23.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==4.2.1
prompt-toolkit==3.0.42
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
pyzmq==26.0.2
requests==2.31.0
rsa==4.9
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
stack-data==0.6.2
starlette==0.37.2
tenacity==8.2.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
wheel==0.43.0
zipp==3.17.0
